+++ .dockerignore
@@ -2,9 +2,9 @@
 !/api
 !/cmd
 !/config
+!/internal
 !/test
 !go.mod
 !go.sum
-!server.go
 !service-account-file.json
 !sonar-project.properties
