+++ .github/workflows/continuous_integration.yml
@@ -76,4 +76,4 @@ jobs:
         with:
           go-version: "^1.19"
       - name: Checking if build
-        run: go build -o ./build/server server.go
+        run: go build -o ./build/server ./cmd/http_server.go
