+++ Dockerfile
@@ -3,7 +3,7 @@ RUN apk --update add build-base
 WORKDIR /app
 COPY . .
 RUN go mod download
-RUN go build -o ./build/server server.go
+RUN go build -o ./build/server ./cmd/http_server.go
 
 FROM alpine:3.16.2 as deploy
 RUN apk --no-cache add ca-certificates
