+++ api/http/routes/email.route.go
@@ -2,7 +2,7 @@
 package routes
 
 import (
-	core "auth-plus-notification/cmd"
+	core "auth-plus-notification/internal"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
