+++ test/unit/usecases/telegram_test.go
@@ -1,8 +1,8 @@
 package test
 
 import (
-	u "auth-plus-notification/cmd/usecases"
-	d "auth-plus-notification/cmd/usecases/driven"
+	u "auth-plus-notification/internal/usecases"
+	d "auth-plus-notification/internal/usecases/driven"
 	t "auth-plus-notification/test/mocks"
 
 	"errors"
